fix: enforce A3 revision bundle contract
